silant_service/requirements.txt
Django==5.2.1
djangorestframework==3.15.1
django-allauth==0.61.1
